+++ src/util/index.ts
@@ -24,3 +24,11 @@ export const fillOverDaysData = (data: Array<{ date: string, count: number }>, o
 
   return res
 }
+
+export function genSQLLatestDayRange(latest: number) {
+  return new Array(latest)
+    .fill(0)
+    .map((_, i) => `SELECT DATE_ADD(CURDATE(), INTERVAL -${i} DAY) AS date`)
+    .reverse()
+    .join(' UNION ALL ')
+}