+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 ### updates
 
 - admin 不考虑做响应式，只提供 sidebar 的展开/收起
-
-
+- 添加 `useBreakPoint`，用于后续前台页面的响应式
+- 完成 `/admin/page` 的部分表单、表格逻辑，分页查询部分需要完善
+- `/admin/dashboard` 添加 `@ant-design/charts` 绘制图表
 
 
 ### bugs
@@ -14,3 +15,6 @@
 
 
 
+### todos
+
+- `/admin/post` 页面需要添加表单筛选查询，并且完善分页参数